canvas: Move contracts config to own file
